added gradle build file and modified manifest
the modified manifest is necessary depending on how the libraryy
is imported into the hosting application.
In the case of android-sqrl, the activity is defined separately
in the main application AndroidManifest.xml, thus the library manifest
should not define the activity on its own (at least not when using
gradle which attempts to merge manifests).